:recycle: moved PersonMixinTestCase into uid.tests.mixins
