+++ django-data/image/uid/tests/__init__.py
@@ -6,6 +6,9 @@ Created on Wed Feb 14 10:33:34 2018
 @author: Paolo Cozzi <paolo.cozzi@ptp.it>
 """
 
-from .mixins import DataSourceMixinTestCase, FileReaderMixinTestCase
+from .mixins import (
+    DataSourceMixinTestCase, FileReaderMixinTestCase, PersonMixinTestCase)
 
-__all__ = ["DataSourceMixinTestCase", "FileReaderMixinTestCase"]
+__all__ = [
+    "DataSourceMixinTestCase", "FileReaderMixinTestCase",
+    "PersonMixinTestCase"]
